backend/requirements.txt
asgiref==3.7.2
Django==3.2.23
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-js-asset==2.2.0
djangorestframework==3.14.0
pillow==10.2.0
pytz==2023.3.post1
sqlparse==0.4.4
